feat(split_all_types): split file with classes and struct
