+++ ClassSplitter.Tests/SmallFileWithStruct.cs
@@ -1,0 +1,12 @@
+namespace ClassSplitter.Tests
+{
+    public struct SmallFileStruct    { }
+
+    public static class SmallFileBrotherClass 
+    {
+        public static string SayHelloToAnother(string friend)
+        {
+            return "Hello, my dear" + friend;
+        }
+    }
+}
